+++ packages/react/src/index.ts
@@ -29,6 +29,7 @@ export * from "./slots";
 export * from "./inventoryGrid";
 export * from "./panels";
 export * from "./characterSheet";
+export * from "./talentTree";
 export * from "./hudTheme";
 export * from "./hudThemePreview";
 export * from "./hudFrame";